Struct storage_device::BlockBounds

pub struct BlockBounds {
    pub range: Range<usize>,
    pub first_block_offset: usize,
    pub last_block_offset: usize,
}

Block-wise bounds information for a data transfer (read or write) that has been calculated from a byte-wise offset and buffer length.

See the block_bounds() method.

Fields
range: Range<usize>

A Range from the first block to the last block of the transfer.

first_block_offset: usize

The offset into the first block (beginning bound) where the transfer starts.

last_block_offset: usize

The offset into the last block (ending bound) where the transfer ends.

Methods

impl BlockBounds

pub fn block_count(&self) -> usize

The total number of blocks to be transferred, i.e., last block - first block.

pub fn is_first_block_aligned(&self) -> bool

Returns true if the first block of the transfer is aligned to a block boundary.

pub fn is_last_block_aligned(&self) -> bool

Returns true if the last block of the transfer is aligned to a block boundary.